Fix test build error
